Refactoring (Frontend): Standardized frontend refactoring, incorporating code contracts.
