+++ package.json
@@ -9,6 +9,10 @@
     "ios": "tauri ios",
     "build": "pnpm run tauri:build",
     "dev": "pnpm run tauri:dev",
+    "check": "pnpm run check:frontend && pnpm run check:types",
+    "check:frontend": "node scripts/check-frontend-guardrails.mjs",
+    "check:frontend:baseline": "node scripts/check-frontend-guardrails.mjs --update-baseline",
+    "check:types": "tsc -p tsconfig.host.json",
     "web:build": "webpack --config webpack.config.js",
     "web:dev": "webpack --config webpack.config.js --watch",
     "tauri:dev": "tauri dev",
@@ -58,6 +62,7 @@
     "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.1",
     "stream-browserify": "^3.0.0",
+    "typescript": "^5.5.4",
     "util": "^0.12.5",
     "webpack-cli": "^5.1.4"
   },
