+++ scripts/guardrails/frontend-lines-baseline.json
@@ -1,0 +1,15 @@
+{
+  "version": 1,
+  "createdAt": "2026-03-11",
+  "maxLinesByPath": {
+    "src/tauri/main/bootstrap.js": 388,
+    "src/tauri/main/context.js": 2,
+    "src/tauri/main/routes/ai-routes.js": 945,
+    "src/tauri/main/routes/chat-routes.js": 646
+  },
+  "notes": [
+    "这些文件是宿主层的关键聚合点：Phase 1 起禁止继续增长。",
+    "如需新增能力，请优先新增小文件（<= 500 行）并通过组合方式接入；不要继续往这些文件堆逻辑。",
+    "当完成拆分重构后，可更新本基线以反映新的、更小的文件布局。"
+  ]
+}
