+++ src/tauri/main/services/assets/asset-service.js
@@ -1,0 +1,120 @@
+// @ts-check
+
+import { joinPath, sanitizeRelativePath } from '../../kernel/path-utils.js';
+
+/**
+ * @typedef {import('../../context/types.js').ConvertFileSrcFn} ConvertFileSrcFn
+ * @typedef {import('../../context/types.js').UserDirectories} UserDirectories
+ * @typedef {import('../../context/types.js').ThumbnailRouteSpec} ThumbnailRouteSpec
+ */
+
+/**
+ * @param {{
+ *   convertFileSrc: ConvertFileSrcFn;
+ *   getUserDirectories: () => UserDirectories | null;
+ *   thumbnailRouteTypes: ReadonlySet<string>;
+ * }} deps
+ */
+export function createAssetService({ convertFileSrc, getUserDirectories, thumbnailRouteTypes }) {
+    /**
+     * @param {string} type
+     * @param {string} file
+     * @param {{ cacheBust?: string | number | null; animated?: boolean } | undefined} options
+     */
+    function buildThumbnailRouteUrl(type, file, { cacheBust = null, animated = false } = {}) {
+        const normalizedType = String(type || '').trim().toLowerCase();
+        const normalizedFile = String(file || '');
+        const searchParams = new URLSearchParams({
+            type: normalizedType,
+            file: normalizedFile,
+        });
+
+        if (animated) {
+            searchParams.set('animated', 'true');
+        }
+
+        if (cacheBust !== null && cacheBust !== undefined && cacheBust !== '') {
+            searchParams.set('t', String(cacheBust));
+        }
+
+        return `/thumbnail?${searchParams.toString()}`;
+    }
+
+    /**
+     * @param {unknown} rawUrl
+     * @returns {ThumbnailRouteSpec | null}
+     */
+    function parseThumbnailRouteUrl(rawUrl) {
+        try {
+            const url = new URL(String(rawUrl || ''), window.location.origin);
+            if (url.pathname !== '/thumbnail') {
+                return null;
+            }
+
+            const type = String(url.searchParams.get('type') || '').trim().toLowerCase();
+            const file = String(url.searchParams.get('file') || '').trim();
+            if (!type || !file || !thumbnailRouteTypes.has(type)) {
+                return null;
+            }
+
+            const animated = String(url.searchParams.get('animated') || '').toLowerCase() === 'true';
+            const cacheBust = String(url.searchParams.get('t') || '').trim();
+            return {
+                type,
+                file,
+                animated,
+                cacheBust,
+            };
+        } catch {
+            return null;
+        }
+    }
+
+    /**
+     * @param {string} type
+     * @param {string} file
+     */
+    function resolveAssetPath(type, file) {
+        const userDirectories = getUserDirectories();
+        if (!userDirectories || !file) {
+            return null;
+        }
+
+        const cleanFile = sanitizeRelativePath(file);
+        if (!cleanFile) {
+            return null;
+        }
+
+        switch (type) {
+            case 'avatar':
+                return joinPath(userDirectories.characters, cleanFile);
+            case 'persona':
+                return joinPath(userDirectories.avatars, cleanFile);
+            case 'bg':
+                return joinPath(userDirectories.backgrounds, cleanFile);
+            default:
+                return null;
+        }
+    }
+
+    /**
+     * @param {string} filePath
+     */
+    function toAssetUrl(filePath) {
+        if (!filePath) {
+            return null;
+        }
+
+        // In Tauri WebView, file:// resources are generally blocked for img/src.
+        // Always use the asset protocol URL generated by convertFileSrc.
+        return convertFileSrc(filePath, 'asset');
+    }
+
+    return {
+        buildThumbnailRouteUrl,
+        parseThumbnailRouteUrl,
+        resolveAssetPath,
+        toAssetUrl,
+    };
+}
+
