+++ tsconfig.host.json
@@ -1,0 +1,30 @@
+{
+  "compilerOptions": {
+    "target": "ES2020",
+    "module": "ES2020",
+    "moduleResolution": "Bundler",
+    "lib": ["ES2020", "DOM", "DOM.Iterable"],
+
+    "allowJs": true,
+    "checkJs": false,
+    "noEmit": true,
+
+    "strict": true,
+    "exactOptionalPropertyTypes": true,
+    "noUncheckedIndexedAccess": true,
+    "useUnknownInCatchVariables": true,
+
+    "skipLibCheck": true,
+    "types": []
+  },
+  "include": [
+    "src/types.d.ts",
+    "src/tauri/main/context/**/*.js",
+    "src/tauri/main/kernel/**/*.js",
+    "src/tauri/main/services/**/*.js",
+    "src/tauri/main/ports/**/*.js",
+    "src/tauri/main/adapters/**/*.js",
+    "src/tauri/main/context.js"
+  ]
+}
+
